+++ app/device/device_exp_ultrasound_settings.h
@@ -22,8 +22,8 @@ class DeviceUltrasound {
  public:
   DeviceUltrasound();
   DeviceUltrasound(int32_t exp_clipping_enable,
-                   int32_t exp_clip_time_duration,
-                   int32_t exp_clip_time_paused,
+                   int64_t exp_clip_time_duration,
+                   int64_t exp_clip_time_paused,
                    int64_t exp_max_cycle_count,
                    int32_t exp_max_cycle_power,
                    int32_t exp_frequency_fluctuations_range);
@@ -34,9 +34,9 @@ class DeviceUltrasound {
   int32_t exp_clipping_enable_;
 
   /// @brief  exp clipping time duration.
-  int32_t exp_clip_time_duration_;
+  int64_t exp_clip_time_duration_;
   /// @brief  exp clipping time paused.
-  int32_t exp_clip_time_paused_;
+  int64_t exp_clip_time_paused_;
 
   /// @brief  exp max cycle count.
   int64_t exp_max_cycle_count_;
